backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.16
psycopg[binary]>=3.2,<3.3
openai>=1.50,<2
pypdf>=5.0,<6
python-docx>=1.1,<1.2
drf-spectacular>=0.27,<0.28
